Add files via upload
The main objective is to predict the price of the used cars of the series of year ranging between 2010- 2020. The dataset has above 100k rows and 8 columns. In this I have performed the exploratory data analysis of various plots and also performed various machine learning techniques to predict the best price for a car.